+++ packages/expo-app-metrics/CHANGELOG.md
@@ -11,3 +11,5 @@
 ### 🐛 Bug fixes
 
 ### 💡 Others
+
+- Add private `getForegroundSession` ([#46657](https://github.com/expo/expo/pull/46657) by [@Ubax](https://github.com/Ubax))